Refactor team image thumbnail generation
This will allow is to actually use the thumbnails in a clean way.
As thumbnails are stored in their own directory now, it also fixes a
potential vulnerability where you could overwrite thumbnails by
registering teams with names ending in '_100x100.png'. (Nice bug BTW,
might use this for a CTF challenge).