Web: Do not overwrite default QuerySet of GameControl
Creating an object on queries led to strange errors when running test
cases: Django's testing facilities store a serialized copy of the
initial (presumably empty) database, in order to restore it after every
test. Except that it wasn't empty if a GameControl object gets created
during serialization. So this would lead to multiple GameControl objects
in the database if another one is created from a fixture.

We could still use a queryset and just don't overwrite the default, but
a simple classmethod seems better suited.