Web: Prevent deletion of Teams after competition has started
Theoretically, there can be cases where registration is still open
(meaning Teams can be edited) after the competition has begun. This is
usually not much of a problem, but deleting a Team in that situation
will technically break the Materialized View for the scoreboard and
semantically scoring as a whole.

Closes: https://github.com/fausecteam/ctf-gameserver/issues/12