Controller: Improve handling of the competition's end
More of a pathological case, but we don't want to increase the Tick at
the exact moment of the competition's end. (Start is included in the
duration, end is not.)

This is hard to test for and the change to the testcase will never
catch it in practice, but let's make our intentions more clear.

Co-authored-by: Simon Ruderich <simon@ruderich.org>