Checker: Fix broken checker script in test case
The call to `get_flag()` used the wrong tick, causing an exception.
This lead to the script getting killed because of a communication error.

This was hidden most of the time because of a race condition:
The kill instruction from above was not executed before
test_unfinished() checked whether the process was still running and
before the script got terminated as part of the test itself.

Co-authored-by: Simon Ruderich <simon@ruderich.org>